feat: add solution for modules: esm task
